refactor actions - new hook because cant use useAddNewCrd, out of scope
